Dashboard refactoring
