+++ src/App.js
@@ -8,7 +8,7 @@ import ProtectedRoute from "./components/Auth/ProtectedRoute"
 import AppForm from "./components/SubmissionForm/AppForm"
 import { AppsGallery } from "./components/AppCard"
 import Home from "./components/LandingPage/Home"
-import { NavBar, Header, Footer } from "./components/reusable"
+import { NavBar, Footer } from "./components/reusable"
 
 import { MuiThemeProvider, createMuiTheme } from "@material-ui/core"
 import "./App.scss"
@@ -56,12 +56,10 @@ const App = ({ loginUser }) => {
       <MuiThemeProvider theme={theme}>
         <div className="App">
           <NavBar />
-          <Header />
-
           <Route exact path="/" component={Home} />
           <ProtectedRoute exact path="/app-form" component={AppForm} />
           <Route path="/apps" exact component={AppsGallery} />
-          <ProtectedRoute path="/dashboard" exact component={Dashboard} />
+          <ProtectedRoute path="/dashboard" component={Dashboard} />
           <ProtectedRoute
             path="/pending-apps"
             exact
