+++ src/components/Dashboard/Dashboard.js
@@ -1,6 +1,10 @@
 import React from 'react';
+import ProtectedRoute from "../Auth/ProtectedRoute";
+import Profile from "./Profile";
+import Overview from "./Overview";
+import { connect } from "react-redux";
+import useStyles from "./styles"
 import clsx from 'clsx';
-import { makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Drawer from '@material-ui/core/Drawer';
 import AppBar from '@material-ui/core/AppBar';
@@ -8,121 +12,14 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Divider from '@material-ui/core/Divider';
 import IconButton from '@material-ui/core/IconButton';
-import Badge from '@material-ui/core/Badge';
 import Container from '@material-ui/core/Container';
-import Grid from '@material-ui/core/Grid';
-import Paper from '@material-ui/core/Paper';
-// import Link from '@material-ui/core/Link'; ---UNCOMMENT IF ADDING COPYRIGHT FOR MUI
 import MenuIcon from '@material-ui/icons/Menu';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import NotificationsIcon from '@material-ui/icons/Notifications';
 import List from '@material-ui/core/List';
 import MainItemList from './MainItemList';
 import SecondaryItemList from './SecondaryItemList'
 
-// function Copyright() {
-//     return (
-//         <Typography variant="body2" color="textSecondary" align="center">
-//             {'Copyright © '}
-//             <Link color="inherit" href="https://material-ui.com/">
-//                 Your Website
-//       </Link>{' '}
-//             {new Date().getFullYear()}
-//             {'. Built with '}
-//             <Link color="inherit" href="https://material-ui.com/">
-//                 Material-UI.
-//       </Link>
-//         </Typography>
-//     );
-// }
-
-const drawerWidth = 240;
-
-const useStyles = makeStyles(theme => ({
-    root: {
-        display: 'flex',
-    },
-    toolbar: {
-        paddingRight: 24,
-        backgroundColor: "#1a61b0",
-    },
-    toolbarIcon: {
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'flex-end',
-        padding: '0 8px',
-        ...theme.mixins.toolbar,
-    },
-    appBar: {
-        zIndex: theme.zIndex.drawer + 1,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen,
-        }),
-    },
-    appBarShift: {
-        marginLeft: drawerWidth,
-        width: `calc(100% - ${drawerWidth}px)`,
-        transition: theme.transitions.create(['width', 'margin'], {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    },
-    menuButton: {
-        marginRight: 36,
-    },
-    menuButtonHidden: {
-        display: 'none',
-    },
-    title: {
-        flexGrow: 1,
-        color: "white",
-    },
-    drawerPaper: {
-        position: 'relative',
-        whiteSpace: 'nowrap',
-        width: drawerWidth,
-        transition: theme.transitions.create('width', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    },
-    drawerPaperClose: {
-        overflowX: 'hidden',
-        transition: theme.transitions.create('width', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen,
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up('sm')]: {
-            width: theme.spacing(9),
-        },
-    },
-    appBarSpacer: theme.mixins.toolbar,
-    content: {
-        flexGrow: 1,
-        height: '100vh',
-        overflow: 'auto',
-    },
-    container: {
-        paddingTop: theme.spacing(4),
-        paddingBottom: theme.spacing(4),
-    },
-    paper: {
-        padding: theme.spacing(2),
-        display: 'flex',
-        overflow: 'auto',
-        flexDirection: 'column',
-    },
-    fixedHeight: {
-        height: 240,
-    },
-    list: {
-        width: "100%"
-    }
-}));
-
-const Dashboard = () => {
+const Dashboard = ({ user }) => {
 
     const classes = useStyles();
     const [open, setOpen] = React.useState(true);
@@ -132,7 +29,6 @@ const Dashboard = () => {
     const handleDrawerClose = () => {
         setOpen(false);
     };
-    const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
     return (
         <div className={classes.root}>
             <CssBaseline />
@@ -148,13 +44,8 @@ const Dashboard = () => {
                         <MenuIcon />
                     </IconButton>
                     <Typography component="h1" variant="h4" noWrap className={classes.title}>
-                        Student Dashboard
+                        {user.first_name ? `${user.first_name}'s` : "User"} Dashboard
              </Typography>
-                    <IconButton color="inherit">
-                        <Badge badgeContent={4} color="secondary">
-                            <NotificationsIcon />
-                        </Badge>
-                    </IconButton>
                 </Toolbar>
             </AppBar>
             <Drawer
@@ -171,7 +62,6 @@ const Dashboard = () => {
                 </div>
                 <Divider />
 
-                {/* List of Hamburger Navigation */}
                 <List className={classes.list}>
                     <MainItemList />
                 </List>
@@ -184,29 +74,21 @@ const Dashboard = () => {
             <main className={classes.content}>
                 <div className={classes.appBarSpacer} />
                 <Container maxWidth="lg" className={classes.container}>
-                    <h1>Overview</h1>
-                    <Grid container spacing={3}>
-                        {/* Chart */}
-                        <Grid item xs={12} md={8} lg={9}>
-                            <Paper className={fixedHeightPaper}>
-                            </Paper>
-                        </Grid>
-                        {/* Recent Deposits */}
-                        <Grid item xs={12} md={4} lg={3}>
-                            <Paper className={fixedHeightPaper}>
-                            </Paper>
-                        </Grid>
-                        {/* Recent Orders */}
-                        <Grid item xs={12}>
-                            <Paper className={classes.paper}>
-                            </Paper>
-                        </Grid>
-                    </Grid>
+                    <ProtectedRoute exact path="/dashboard" component={Overview} />
+                    <ProtectedRoute path="/dashboard/profile" component={Profile} />
                 </Container>
-                {/* <Copyright /> */}
             </main>
         </div>
     );
 }
 
-export default Dashboard
+const mapStateToProps = ({ usersReducer }) => {
+    return {
+        ...usersReducer,
+    }
+}
+
+export default connect(
+    mapStateToProps,
+    {}
+)(Dashboard)
