+++ src/components/Dashboard/Overview.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import Grid from '@material-ui/core/Grid';
+import Paper from '@material-ui/core/Paper';
+import useStyles from "./styles";
+import clsx from 'clsx';
+
+const Overview = () => {
+    const classes = useStyles();
+    const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
+    return (
+        <>
+            <h1>Overview</h1>
+            <Grid container spacing={3}>
+                <Grid item xs={12} md={8} lg={9}>
+                    <Paper className={fixedHeightPaper}>
+                    </Paper>
+                </Grid>
+                <Grid item xs={12} md={4} lg={3}>
+                    <Paper className={fixedHeightPaper}>
+                    </Paper>
+                </Grid>
+                <Grid item xs={12}>
+                    <Paper className={classes.paper}>
+                    </Paper>
+                </Grid>
+            </Grid>
+        </>
+    )
+}
+
+export default Overview;
