+++ src/components/Dashboard/Profile.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+
+const Profile = () => {
+    return (
+        <div>
+            Hi
+        </div>
+    )
+}
+
+
+export default Profile;
